move minmaxAngleInTriangle to meshutilities
