+++ tests/test-equilateral-stats/test_equilateral_stats.cpp
@@ -1,4 +1,5 @@
 #include "cube.h"
+#include "meshutilities.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
@@ -24,67 +25,11 @@ using testing::FloatEq;
 // we have two main case of non equilaterality
 // the pointy one : like an arrow -> one angle near to 0.
 // the flat triangle: one angle near pi
+// so basically we have minmax.second - minmax.first that is a good metrics
+// to know how much we are not equilateral
 
 
 
-
-template <typename MeshType, typename FaceHandle>
-auto minmaxAngleInTriangle( MeshType const& mesh, FaceHandle const& faceHandle )
-{
-    // Here we assume that a faces contains only 3 vertices
-    auto firstPointIt  = mesh.cfv_iter( faceHandle );
-    auto secondPointIt = std::next( firstPointIt );
-    auto thirdPointIt  = std::next( secondPointIt );
-
-
-    // this should be a function
-    auto vecFromPoint = [&mesh]( auto const& originIt, auto const& destinationIt ) {
-        return mesh.point( *destinationIt ) - mesh.point( *originIt );
-    };
-
-    auto edge0 = vecFromPoint( firstPointIt, secondPointIt );
-    auto edge1 = vecFromPoint( secondPointIt, thirdPointIt );
-    auto edge2 = vecFromPoint( thirdPointIt, firstPointIt );
-
-    edge0.normalize();
-    edge1.normalize();
-    edge2.normalize();
-
-    // ok so we have edge0, edge1,edge2 circulating the triangle
-    // the first angle is between edge0 and -edge2
-    // the second angle is between -edge0 and edge1
-    // the third angle is between -edge1 and edge2
-
-    auto firstCosTheta  = dot( edge0, -edge2 );
-    auto secondCosTheta = dot( -edge0, edge1 );
-    auto thirdCosTheta  = dot( -edge1, edge2 );
-
-    auto restrictAcos = []( auto const& value ) {
-        if ( value > -1.f || value < 1.f )
-        {
-            return value;
-        }
-        else if ( value >= 1.f )
-        {
-            return 1.f;
-        }
-        else
-        {
-            return -1.f;
-        }
-    };
-
-    firstCosTheta  = restrictAcos( firstCosTheta );
-    secondCosTheta = restrictAcos( secondCosTheta );
-    thirdCosTheta  = restrictAcos( thirdCosTheta );
-
-    std::array<float, 3> angles{std::acos( firstCosTheta ), std::acos( secondCosTheta ),
-                                std::acos( thirdCosTheta )};
-
-    auto minmax = std::minmax_element( std::begin( angles ), std::end( angles ) );
-
-    return std::make_pair( *minmax.first, *minmax.second );
-}
 
 TEST( anIsoceleRectangularTriangle, haveAngleOfPiOver4AndPiOver2 )
 {
